+++ ci/docker-compose-exit.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+docker-compose up --exit-code-from "$1" "$1"