+++ codex/librarian/crond.py
@@ -19,23 +19,22 @@ class Crond(Thread):
     COND = Condition()
     run_thread = True
     WAIT_INTERVAL = 60 * 5  # Run cron every 5 minutes
+    SHUTDOWN_TIMEOUT = 5
 
-    def worker(self, cond):
+    def run(self):
         """Watch a path and log the events."""
         LOG.info("Started cron")
-        with cond:
+        with self.COND:
             while self.run_thread:
                 QUEUE.put(ScannerCronTask(sleep=0))
                 QUEUE.put(WatcherCronTask(sleep=0))
                 QUEUE.put(UpdateCronTask(sleep=0, force=False))
-                cond.wait(timeout=self.WAIT_INTERVAL)
+                self.COND.wait(timeout=self.WAIT_INTERVAL)
         LOG.info("Stopped cron.")
 
     def __init__(self):
         """Intialize this thread with the worker."""
-        super().__init__(
-            target=self.worker, name="crond", args=(self.COND,), daemon=True
-        )
+        super().__init__(name="crond", daemon=True)
 
     def stop(self):
         """Stop the cron thread."""