+++ codex/librarian/watcherd.py
@@ -126,6 +126,8 @@ class CodexLibraryEventHandler(FileSystemEventHandler):
 class Uatu(Observer):
     """Watch over librarys from the blue area of the moon."""
 
+    SHUTDOWN_TIMEOUT = 5
+
     def __init__(self, *args, **kwargs):
         """Intialize pk to watches dict."""
         super().__init__(*args, **kwargs)
@@ -186,5 +188,5 @@ class Uatu(Observer):
     def unschedule_all(self):
         """Unschedule all watches."""
         super().unschedule_all()
-        self._pk_watches = {}
+        self._pk_watches = dict()
         LOG.info("Stopped watching all libraries")