+++ frontend/src/components/browser-filter-select.vue
@@ -9,7 +9,7 @@
     hide-details="auto"
     ripple
     :menu-props="{
-      maxHeight: '90vh',
+      maxHeight: '80vh',
       overflowY: false,
       contentClass: filterMenuClass,
     }"