+++ NEWS.md
@@ -1,5 +1,10 @@
 # Codex News
 
+## v0.6.5
+
+- Build Docker images for amd64 & arm64
+- Dark admin pages
+
 ## v0.6.4
 
 - Fix reader bug that only displayed first page